Major update on the logon scripts
Moved the list of elements to vars.php because it will be used in many
places.
also cleaned up navbar a bit to reflect the new vars.php file.
finaly updated the javascript, again for the login and using the dynamic
nature or the vars.php file.
Might need a bit of debugging but looks ok for the moment.