Beginning of the allow users to update the profil
Added the main PHP and js stuff for the users to update.
I'll have to rack my brain and see if I have to go with the admin
account of if it's easy to update AD security.

For the moment, users can update phone numbers but description update
fails (and with no warning)